requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
pillow
streamlit